Support composable Host Metadata
Take advantage of Salt's pillar merging to compose host metadata; in
this way, we can apply multiple roles to a node and have the Zabbix
agent report the combination of applied roles to the Zabbix server. In
turn, we can automatically apply role-specific templates through the
Zabbix server's host registration process.

The following pillar data structures are all valid:

- zabbix-agent: # Not mergable
    hostmetadata: my-host-metadata # Not mergable

- zabbix-agent: # Not mergable; Zabbix cannot merge lists
    hostmetadata:
      - metadata component 1
      - metadata component 2
      - metadata component 3

- zabbix-agent: # Supports pillar merging
    hostmetadata:
      metadata component 1:
      metadata component 2:
      metadata component 3: